Fixing file and directory nomenclature
