+++ src/App.tsx
@@ -343,15 +343,15 @@ import TestNicThreeModuleResetTestPassed from './pages/SystemDiagnostics/FortyTw
 import TestPrcOneModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleOne/TestPrcOneModuleResetTestOne';
 import TestPrcOneModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleOne/TestPrcOneModuleResetTestTwo';
 import TestPrcOneModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleOne/TestPrcOneModuleResetTestPassed';
-import TestPrcTwoModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyTwoProcessorModuleOne/TestPrcTwoModuleResetTestOne';
-import TestPrcTwoModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyTwoProcessorModuleOne/TestPrcTwoModuleResetTestTwo';
-import TestPrcTwoModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyTwoProcessorModuleOne/TestPrcTwoModuleResetTestPassed';
-import TestPrcThreeModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyThreeProcessorModuleOne/TestPrcThreeModuleResetTestOne';
-import TestPrcThreeModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyThreeProcessorModuleOne/TestPrcThreeModuleResetTestTwo';
-import TestPrcThreeModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyThreeProcessorModuleOne/TestPrcThreeModuleResetTestPassed';
-import TestPrcFourModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyFourProcessorModuleOne/TestPrcFourModuleResetTestOne';
-import TestPrcFourModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyFourProcessorModuleOne/TestPrcFourModuleResetTestTwo';
-import TestPrcFourModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyFourProcessorModuleOne/TestPrcFourModuleResetTestPassed';
+import TestPrcTwoModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleTwo/TestPrcTwoModuleResetTestOne';
+import TestPrcTwoModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleTwo/TestPrcTwoModuleResetTestTwo';
+import TestPrcTwoModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleTwo/TestPrcTwoModuleResetTestPassed';
+import TestPrcThreeModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleThree/TestPrcThreeModuleResetTestOne';
+import TestPrcThreeModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleThree/TestPrcThreeModuleResetTestTwo';
+import TestPrcThreeModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleThree/TestPrcThreeModuleResetTestPassed';
+import TestPrcFourModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleFour/TestPrcFourModuleResetTestOne';
+import TestPrcFourModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleFour/TestPrcFourModuleResetTestTwo';
+import TestPrcFourModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleFour/TestPrcFourModuleResetTestPassed';
 import TestPrcFiveModuleResetTestOne from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyProcessorModuleFive/TestPrcFiveModuleResetTestOne';
 import TestPrcFiveModuleResetTestTwo from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyProcessorModuleFive/TestPrcFiveModuleResetTestTwo';
 import TestPrcFiveModuleResetTestPassed from './pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyProcessorModuleFive/TestPrcFiveModuleResetTestPassed';
@@ -684,15 +684,15 @@ function App() {
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprconemoduleresettestone" element={<TestPrcOneModuleResetTestOne />} />
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprconemoduleresettesttwo" element={<TestPrcOneModuleResetTestTwo />} />
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprconemoduleresettestpassed" element={<TestPrcOneModuleResetTestPassed />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprctwomoduleresettestone" element={<TestPrcTwoModuleResetTestOne />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprctwomoduleresettesttwo" element={<TestPrcTwoModuleResetTestTwo />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprctwomoduleresettestpassed" element={<TestPrcTwoModuleResetTestPassed />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcthreemoduleresettestone" element={<TestPrcThreeModuleResetTestOne />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcthreemoduleresettesttwo" element={<TestPrcThreeModuleResetTestTwo />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcthreemoduleresettestpassed" element={<TestPrcThreeModuleResetTestPassed />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcfourmoduleresettestone" element={<TestPrcFourModuleResetTestOne />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcfourmoduleresettesttwo" element={<TestPrcFourModuleResetTestTwo />} />
-        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduleone/testprcfourmoduleresettestpassed" element={<TestPrcFourModuleResetTestPassed />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduletwo/testprctwomoduleresettestone" element={<TestPrcTwoModuleResetTestOne />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduletwo/testprctwomoduleresettesttwo" element={<TestPrcTwoModuleResetTestTwo />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormoduletwo/testprctwomoduleresettestpassed" element={<TestPrcTwoModuleResetTestPassed />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulethree/testprcthreemoduleresettestone" element={<TestPrcThreeModuleResetTestOne />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulethree/testprcthreemoduleresettesttwo" element={<TestPrcThreeModuleResetTestTwo />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulethree/testprcthreemoduleresettestpassed" element={<TestPrcThreeModuleResetTestPassed />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulefour/testprcfourmoduleresettestone" element={<TestPrcFourModuleResetTestOne />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulefour/testprcfourmoduleresettesttwo" element={<TestPrcFourModuleResetTestTwo />} />
+        <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyoneprocessormodulefour/testprcfourmoduleresettestpassed" element={<TestPrcFourModuleResetTestPassed />} />
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyprocessormodulefive/testprcfivermoduleresettestone" element={<TestPrcFiveModuleResetTestOne />} />
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyprocessormodulefive/testprcfivermoduleresettesttwo" element={<TestPrcFiveModuleResetTestTwo />} />
         <Route path="/systemdiagnostics/fortytwointegratedmodularavionics/twentyprocessormodulefive/testprcfivermoduleresettestpassed" element={<TestPrcFiveModuleResetTestPassed />} />
