+++ src/pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleFour/DataFooter.module.scss
