+++ src/pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleThree/DataFooter.module.scss
