+++ src/pages/SystemDiagnostics/FortyTwoIntegratedModularAvionics/TwentyOneProcessorModuleTwo/DataFooter.module.scss
